Refactor SVEditorControls
Now there is a base class called eh_CustomEditorControls that can
handle any custom resources, and SVEditorControls extend from it

Shared Inspector was also refactored and now can deal with any
Resource type, so it can be used anywhere, independently of
Shared Variables or even eh_CustomEditorControls.

This is in preparation to move some things out of this plugin and
into a future eh_Utilities plugin.